Move test models into internal to prevent export
