+++ fixtures/case.json
@@ -1,8 +1,8 @@
 {
-    "$ref": "#/definitions/jsonschema_test.ExampleCase",
+    "$ref": "#/definitions/testmodels.ExampleCase",
     "$schema": "http://json-schema.org/draft-07/schema#",
     "definitions": {
-        "jsonschema_test.BoolPayload": {
+        "testmodels.BoolPayload": {
             "additionalProperties": false,
             "properties": {
                 "payload": {
@@ -14,7 +14,7 @@
             ],
             "type": "object"
         },
-        "jsonschema_test.ExampleCase": {
+        "testmodels.ExampleCase": {
             "additionalProperties": false,
             "oneOf": [
                 {
@@ -37,7 +37,7 @@
                         }
                     },
                     "then": {
-                        "$ref": "#/definitions/jsonschema_test.BoolPayload",
+                        "$ref": "#/definitions/testmodels.BoolPayload",
                         "$schema": "http://json-schema.org/draft-07/schema#"
                     }
                 },
@@ -61,7 +61,7 @@
                         }
                     },
                     "then": {
-                        "$ref": "#/definitions/jsonschema_test.IntPayload",
+                        "$ref": "#/definitions/testmodels.IntPayload",
                         "$schema": "http://json-schema.org/draft-07/schema#"
                     }
                 },
@@ -85,7 +85,7 @@
                         }
                     },
                     "then": {
-                        "$ref": "#/definitions/jsonschema_test.StringPayload",
+                        "$ref": "#/definitions/testmodels.StringPayload",
                         "$schema": "http://json-schema.org/draft-07/schema#"
                     }
                 }
@@ -97,7 +97,7 @@
             },
             "type": "object"
         },
-        "jsonschema_test.IntPayload": {
+        "testmodels.IntPayload": {
             "additionalProperties": false,
             "properties": {
                 "payload": {
@@ -109,7 +109,7 @@
             ],
             "type": "object"
         },
-        "jsonschema_test.StringPayload": {
+        "testmodels.StringPayload": {
             "additionalProperties": false,
             "properties": {
                 "payload": {
