+++ fixtures/defaults_expanded_toplevel.json
@@ -2,7 +2,7 @@
   "$schema": "http://json-schema.org/draft-07/schema#",
   "additionalProperties": false,
   "definitions": {
-    "jsonschema_test.GrandfatherType": {
+    "testmodels.GrandfatherType": {
       "additionalProperties": false,
       "properties": {
         "family_name": {
@@ -12,7 +12,7 @@
       "required": ["family_name"],
       "type": "object"
     },
-    "jsonschema_test.TestUser": {
+    "testmodels.TestUser": {
       "additionalProperties": false,
       "properties": {
         "SomeUntaggedBaseProperty": {
@@ -50,7 +50,7 @@
           "type": "array"
         },
         "grand": {
-          "$ref": "#/definitions/jsonschema_test.GrandfatherType",
+          "$ref": "#/definitions/testmodels.GrandfatherType",
           "$schema": "http://json-schema.org/draft-07/schema#"
         },
         "id": {
@@ -152,7 +152,7 @@
       "type": "array"
     },
     "grand": {
-      "$ref": "#/definitions/jsonschema_test.GrandfatherType",
+      "$ref": "#/definitions/testmodels.GrandfatherType",
       "$schema": "http://json-schema.org/draft-07/schema#"
     },
     "id": {
