+++ fixtures/test_versioned_packages.json
@@ -1,6 +1,6 @@
 {
   "definitions": {
-    "jsonschema_test.DeveloperPackage": {
+    "testmodels.DeveloperPackage": {
       "additionalProperties": false,
       "properties": {
         "experience": {
@@ -31,7 +31,7 @@
       "required": ["experience", "language", "hardware", "hardware1"],
       "type": "object"
     },
-    "jsonschema_test.TesterPackage": {
+    "testmodels.TesterPackage": {
       "additionalProperties": false,
       "properties": {
         "experience": {
@@ -110,10 +110,10 @@
     }
   },
   "oneOf": [{
-    "$ref": "#/definitions/jsonschema_test.TesterPackage",
+    "$ref": "#/definitions/testmodels.TesterPackage",
     "$schema": "http://json-schema.org/draft-07/schema#"
   }, {
-    "$ref": "#/definitions/jsonschema_test.DeveloperPackage",
+    "$ref": "#/definitions/testmodels.DeveloperPackage",
     "$schema": "http://json-schema.org/draft-07/schema#"
   }]
 }
