+++ internal/testmodels/min_max_items.go
@@ -1,4 +1,4 @@
-package jsonschema_test
+package testmodels
 
 type SliceTestType []string
