+++ reflect_test.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"encoding/json"
 	"github.com/discovery-digital/jsonschema"
+	"github.com/discovery-digital/jsonschema/internal/testmodels"
 	"io/ioutil"
 	"path/filepath"
 	"strings"
@@ -17,15 +18,15 @@ type testSet struct {
 }
 
 var schemaGenerationTests = []testSet{
-	{&jsonschema.Reflector{}, "fixtures/defaults.json", TestUser{}},
-	{&jsonschema.Reflector{AllowAdditionalProperties: true}, "fixtures/allow_additional_props.json", TestUser{}},
-	{&jsonschema.Reflector{RequiredFromJSONSchemaTags: true}, "fixtures/required_from_jsontags.json", TestUser{}},
-	{&jsonschema.Reflector{ExpandedStruct: true}, "fixtures/defaults_expanded_toplevel.json", TestUser{}},
-	{&jsonschema.Reflector{}, "fixtures/test_one_of_default.json", TestUserOneOf{}},
-	{&jsonschema.Reflector{}, "fixtures/test_versioned_packages.json", TestVersionedPackages{}},
-	{&jsonschema.Reflector{}, "fixtures/if_then_else.json", Application{}},
-	{&jsonschema.Reflector{}, "fixtures/case.json", ExampleCase{}},
-	{&jsonschema.Reflector{}, "fixtures/test_min_max_items.json", SliceTestType{}},
+	{&jsonschema.Reflector{}, "fixtures/defaults.json", testmodels.TestUser{}},
+	{&jsonschema.Reflector{AllowAdditionalProperties: true}, "fixtures/allow_additional_props.json", testmodels.TestUser{}},
+	{&jsonschema.Reflector{RequiredFromJSONSchemaTags: true}, "fixtures/required_from_jsontags.json", testmodels.TestUser{}},
+	{&jsonschema.Reflector{ExpandedStruct: true}, "fixtures/defaults_expanded_toplevel.json", testmodels.TestUser{}},
+	{&jsonschema.Reflector{}, "fixtures/test_one_of_default.json", testmodels.TestUserOneOf{}},
+	{&jsonschema.Reflector{}, "fixtures/test_versioned_packages.json", testmodels.TestVersionedPackages{}},
+	{&jsonschema.Reflector{}, "fixtures/if_then_else.json", testmodels.Application{}},
+	{&jsonschema.Reflector{}, "fixtures/case.json", testmodels.ExampleCase{}},
+	{&jsonschema.Reflector{}, "fixtures/test_min_max_items.json", testmodels.SliceTestType{}},
 }
 
 func TestSchemaGeneration(t *testing.T) {
@@ -35,12 +36,12 @@ func TestSchemaGeneration(t *testing.T) {
 }
 func TestOverrides(t *testing.T) {
 	override := jsonschema.GetSchemaTagOverride()
-	override.Set(Hardware{}, "Brand", "enum=microsoft|apple|lenovo|dell")
+	override.Set(testmodels.Hardware{}, "Brand", "enum=microsoft|apple|lenovo|dell")
 
 	test := testSet{
 		reflector: &jsonschema.Reflector{Overrides: override},
 		fixture:   "fixtures/override_jsonschema_tag.json",
-		actual:    TestUserOneOf{},
+		actual:    testmodels.TestUserOneOf{},
 	}
 
 	runTests(t, test)
